Refactor project structure: separate layers
