+++ backend/src/controllers/user.controller.ts
@@ -1,5 +1,5 @@
 import { Request, Response } from 'express';
-import * as userService from './user.service';
+import * as userService from '../services/user.service';
 
 export const createUser = async (req: Request, res: Response) => {
   try {
