+++ backend/src/models/index.ts
@@ -1,0 +1,4 @@
+export * from './user.model';
+export * from './prompt.model';
+export * from './category.model';
+export * from './sub-category.model';
