+++ backend/src/routes/user.routes.ts
@@ -1,5 +1,5 @@
 import { Router } from 'express';
-import * as userController from './user.controller';
+import * as userController from '../controllers/user.controller';
 
 const router = Router();
